requirements.txt
fastapi
uvicorn[standard]

# DB
sqlalchemy
asyncpg
psycopg2-binary

# Auth
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation
pydantic
email-validator

# Env
python-dotenv

# Forms
python-multipart

# Logging (optional but good)
loguru
pydantic-settings

Pillow
python-magic  # For file type detection (optional but recommended)